unit tests from request.rs moved to tests/unit/request.rs
